moved util to core/util
